backend/requirements/base.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Utilities
python-multipart==0.0.6
PyYAML==6.0.1

# email validation 
email-validator==2.1.1